+++ lib/src/db/v1_types.rs
@@ -32,14 +32,6 @@ pub enum SubResourceV1 {
 pub type PropValsV1 = HashMap<String, ValueV1>;
 
 #[derive(Debug, Serialize, Deserialize)]
-#[allow(dead_code)]
-pub struct QueryFilterV1 {
-    pub property: Option<String>,
-    pub value: Option<ValueV1>,
-    pub sort_by: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
 pub struct ResourceV1 {
     propvals: PropValsV1,
     subject: String,